User workers conditionatly when it required

+++ frontend/src/hooks/useSortWorker/useSortWorker.ts
@@ -1,7 +1,8 @@
 import { useEffect, useState } from "react";
 import { SortTypes } from "src/helpers/utils/sortBy/sortBy";
-import SortWorker from "src/helpers/utils/sortWorker.ts?worker";
-import { Value } from "./useForm";
+import { ITEMS_COUNT_TO_USE_WORKER } from "src/constants";
+import sortByFn from "src/helpers/utils/sortBy/sortBy";
+import { Value } from "../useForm";
 
 export const useSortWorker = <T>(
   data: T[],
@@ -13,8 +14,11 @@ export const useSortWorker = <T>(
   const [isWorking, setIsWorking] = useState(false);
 
   useEffect(() => {
-    const worker = new SortWorker();
-    if (data.length && sortBy) {
+    let worker: Worker | null = null;
+
+    const sortWithWorker = async () => {
+      const { default: SortWorker } = await import("./sortWorker?worker");
+      worker = new SortWorker();
       worker.postMessage({
         data,
         sortByField: sortBy,
@@ -33,17 +37,29 @@ export const useSortWorker = <T>(
         setSortedData(data || []);
         setIsWorking(false);
       };
+    };
+
+    if (data.length && sortBy && data.length > ITEMS_COUNT_TO_USE_WORKER) {
+      sortWithWorker();
+    } else if (sortBy) {
+      const sortedData = sortByFn(
+        data,
+        sortBy.toString(),
+        Boolean(isSortedAsc),
+        sortType
+      );
+      setSortedData(sortedData);
     } else {
       setSortedData(data || []);
     }
 
     return () => {
-      worker.terminate();
+      if (worker) {
+        worker.terminate();
+      }
       setIsWorking(false);
     };
   }, [data, sortBy, isSortedAsc, sortType]);
 
   return { sortedData, isWorking };
 };
-
-export default useSortWorker;
